requirements.txt
pip==21.1.2
PyYAML==6.0
setuptools==57.0.0
wheel==0.36.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-telegram-bot==20.0a4
